% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Blocks.R
\name{Blocks}
\alias{Blocks}
\title{Blocks object.}
\usage{
Blocks(NeighborsObj, Neighbors, keysUsed)
}
\arguments{
\item{NeighborsObj}{\code{\link{Neighbors}} object to perform BSN.}

\item{Neighbors}{Data frame with rows as potential matches and
columns for potential match pair IDs. Sorted.}

\item{keysUsed}{List containing blocking variables and sort keys
for each iteration of the algorithm.}
}
\value{
An object of class \code{\link{Blocks}}.
}
\description{
Stores the neighbors and parameters from calling \code{blockedSN}.
}
