% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deDuplicate.R
\name{deDuplicate}
\alias{deDuplicate}
\alias{deDuplicate.default}
\alias{deDuplicate.Scores}
\title{deDuplicate scored neighbors matrix.}
\usage{
deDuplicate(object, thresh, priority, isProband, dateFormat = NULL, requestID)

\method{deDuplicate}{default}(object, thresh, priority, isProband, dateFormat = NULL, requestID)

\method{deDuplicate}{Scores}(object, thresh, priority, isProband, dateFormat = NULL, requestID)
}
\arguments{
\item{object}{\code{\link{Scores}} object containing pre-computed Neighbors and match scores.}

\item{thresh}{Vector of thresholds at which to classify scores as duplicates. These can
either be integers from 1 to 7 (for the intersection score) or numbers between 0 and 1,
representing quantiles (for the greedy match score).}

\item{priority}{A list of structure (var = 'Varx', min = TRUE) with 'Varx'
being a character value corresponding to a column in rawData.
This parameter determines how to sort the duplicates.
If \code{min = TRUE}, then we use the minimum value of 'Varx' for each duplicate entity.
Otherwise, we use the maximum value.}

\item{isProband}{A character value indicating the column indicator column that denotes the proband.}

\item{dateFormat}{Character string of the format of the date. This is only used
if the priority variable is a date. The format should match the formats of
class \code{POSIXlt} used in the \code{base::strptime} function.}

\item{requestID}{Column that has the ID for the family.}
}
\value{
An object of class \code{\link{Duplicates}} containing the duplicate entities
and representatives for each duplicate entity (including singletons without duplicates)..
}
\description{
Outputs the \code{\link{Duplicates}} object, providing the duplicate entities and
the representatives for each duplicate entity.
}
