% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duplicates.R
\name{Duplicates}
\alias{Duplicates}
\title{Duplicates object.}
\usage{
Duplicates(ScoresObj, dupsList, dupsReps, details)
}
\arguments{
\item{ScoresObj}{\code{\link{Scores}} object.}

\item{dupsList}{A list of the duplicate entities. Each duplicate entity is
a vector of request IDs, where the corresponding families are all considered
duplicates of each other.}

\item{dupsReps}{A vector of duplicate ID representatives for each duplicate entity.
The chosen representative is based on the priority variable used in \code{\link{deDuplicate}}.}

\item{details}{Details to be printed regarding the thresholding and ordering}
}
\value{
An object of class \code{\link{Duplicates}}.
}
\description{
Stores the results from calling \code{\link{deDuplicate}}.
}
