% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedyConcordance.R
\name{greedyConcordance}
\alias{greedyConcordance}
\title{Calculate greedy match score.}
\usage{
greedyConcordance(fam1, fam2, keyVars)
}
\arguments{
\item{fam1}{Data frame containing family members 1}

\item{fam2}{Data frame containing familiy members 2}

\item{keyVars}{key variables matarix}
}
\value{
Score value, numeric.
}
\description{
\code{greedyConcordance} scores pedigree members in a sequential greedy fashion
maximizing the match score and eliminating that family member from future
consideration at each iteration.
}
