% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/injectDuplicates.R
\name{injectDuplicates.Neighbors}
\alias{injectDuplicates.Neighbors}
\title{Inject duplicates.}
\usage{
injectDuplicates.Neighbors(orig.data, pDup, errRt)
}
\arguments{
\item{orig.data}{Data}

\item{pDup}{Numeric value. Proportion of dataset number of rows to inject duplicates with.}

\item{errRt}{Parameter of the Poisson distribution for the number of errors in each duplicate.}
}
\value{
An object of class \code{\link{Blocks}} containing the neighbors found
and keys used during the blocked sorted neighbors iteration.
}
\description{
\code{injectDuplicates} returns a data frame with known duplicates added to the rows.
}
