% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Neighbors.R
\name{Neighbors}
\alias{Neighbors}
\title{Neighbors object.}
\usage{
Neighbors(dat, ID, keyVars, keyWt = NULL)
}
\arguments{
\item{dat}{Input data matrix used in the SNIP algorithm.}

\item{ID}{Character or numeric value. Identifying column in dat.
This is the ID variable for which \code{Neighbors} will reference.}

\item{keyVars}{Character or numeric vector. Which variables to
include when generating sort keys?}

\item{keyWt}{Optional vector of weights that determine the probability of selecting
each variable when generating the sort key}
}
\value{
An object of class \code{\link{Neighbors}}.
}
\description{
Stores the input data and parameters for calling \code{blockedSN}.
}
