% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.Scores}
\alias{plot.Scores}
\alias{plot.Duplicates}
\title{Plot Neighbors object generic.}
\usage{
\method{plot}{Scores}(x, type, ...)

\method{plot}{Duplicates}(x, type, ...)
}
\arguments{
\item{x}{code{\link{Neighbors}} or \code{\link{Neighbors}} oject to be summarized.}

\item{type}{Char, How to display the plot (eg. 'box', 'density')}

\item{...}{other arguments to be passed.}
}
\description{
\code{plot} plots the contents of the object.
}
