% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformatDuplicates.R
\name{reformatDuplicates}
\alias{reformatDuplicates}
\alias{reformatDuplicates.default}
\alias{reformatDuplicates.Duplicates}
\title{format deduplicated data back to a list for exporting.}
\usage{
reformatDuplicates(object)

\method{reformatDuplicates}{default}(object)

\method{reformatDuplicates}{Duplicates}(object)
}
\arguments{
\item{object}{\code{\link{Duplicates}} object.}
}
\value{
An object of class list containing data matrices ordered according
to their duplicate status (1=unique, 2=duplicates1, 3=duplicates2, etc.)
}
\description{
format deduplicated data back to a list for exporting.
}
