% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revealTruth.R
\name{revealTruth}
\alias{revealTruth}
\alias{revealTruth.default}
\alias{revealTruth.Duplicates}
\alias{revealTruth.Scores}
\title{Reveals the performance of the deduplication process.}
\usage{
revealTruth(object, thresh, requestID, famID)

\method{revealTruth}{default}(object, thresh, requestID, famID)

\method{revealTruth}{Duplicates}(object, requestID, famID)

\method{revealTruth}{Scores}(object, requestID, famID, thresh)
}
\arguments{
\item{object}{\code{\link{Scores}} object}

\item{thresh}{Vector of thresholds at which to classify scores as duplicates. These can
either be integers from 1 to 7 (for the intersection score) or numbers between 0 and 1,
representing quantiles (for the greedy match score). If the \code{Scores} object has scores
for both the intersection and greedy match scores, then \code{thresh} is a list of 2 vectors,
where the first element is named "intersection" and the second element is named "greedy".}

\item{requestID}{Character value for the column that identifies the apparent family ID}

\item{famID}{Character value for the column that identifies the true family ID}
}
\value{
An object of class \code{\link{Truth}}.
}
\description{
Reveals the performance of the deduplication process.

Reveals
}
