% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleBackground.R
\name{sampleBackground}
\alias{sampleBackground}
\alias{sampleBackground.default}
\alias{sampleBackground.Neighbors}
\alias{sampleBackground.Blocks}
\alias{sampleBackground.Scores}
\title{Populate Neighbors with random pairs.}
\usage{
sampleBackground(object, num)

\method{sampleBackground}{default}(object, num)

\method{sampleBackground}{Neighbors}(object, num)

\method{sampleBackground}{Blocks}(object, num)

\method{sampleBackground}{Scores}(object, num)
}
\arguments{
\item{object}{object containing pre-computed Neighbors from the BSN algorithm.}

\item{num}{Number or proportion of current Neighbors if <1 to populate with.}
}
\value{
An object of class \code{\link{Neighbors}} containing the possible duplicate pairs
sampled at random.
}
\description{
Populate Neighbors with random pairs.
}
