% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreNeighbors.R
\name{scoreNeighbors}
\alias{scoreNeighbors}
\alias{scoreNeighbors.default}
\alias{scoreNeighbors.Blocks}
\alias{scoreNeighbors.Scores}
\title{Calculate match score generic.}
\usage{
scoreNeighbors(object, method = "intersection")

\method{scoreNeighbors}{default}(object, method = "intersection")

\method{scoreNeighbors}{Blocks}(object, method = "intersection")

\method{scoreNeighbors}{Scores}(object, method = "intersection")
}
\arguments{
\item{object}{object containing pre-computed Neighbors from the BSN algorithm.}

\item{method}{Character value.  If intersection, then uses the normalized count of times the match was found. Else uses variable scoring.}
}
\value{
An object of class \code{\link{Scores}} containing the scored neighbor pairs.
}
\description{
\code{scoreNeighbors} returns a \code{\link{Scores}} object with the \code{Neighbors}
slot populated with a match score for each potential duplicate found from the BSN algorithm.
A match score is calculated by averaging the number of string/binary variable matches
consistent between the two records plus a similarity measure for the continuous variables.
}
