% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{set}
\alias{set}
\alias{set.Neighbors}
\alias{set.Blocks}
\alias{set.Scores}
\title{Set Neighbors attribute generic.}
\usage{
set(object, attr, val)

\method{set}{Neighbors}(object, attr, val)

\method{set}{Blocks}(object, attr, val)

\method{set}{Scores}(object, attr, val)
}
\arguments{
\item{object}{object to be updated.}

\item{attr}{Character input. The attribute to update (e.g. 'rawData', 'ID')}

\item{val}{The new values to add or update. For attributes 'Neighbors' and 'keysUsed',
\code{set} will concatenate the values. Otherwise old values will be replaced.}
}
\value{
object of same class (\code{\link{Neighbors}}, \code{\link{Blocks}}, etc.).
}
\description{
\code{set} sets the attribute for the \code{\link{Neighbors}} object.
}
