% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snip.R
\name{snip}
\alias{snip}
\title{Run entire SNIP deduplication algorithm}
\usage{
snip(
  pedigrees,
  requestID,
  isProband,
  keyVars,
  keyVars.male = NULL,
  keyVars.female = NULL,
  keyWt = NULL,
  blockVar = NULL,
  repSN = 1,
  windowSN = 10,
  keyLength = length(keyVars),
  method = "intersection",
  thresh = 1:7,
  priority,
  dateFormat = NULL,
  printRuntime = TRUE,
  seed = NULL
)
}
\arguments{
\item{pedigrees}{Pedigree data to deduplicate.}

\item{requestID}{Column that has the ID for the family.}

\item{isProband}{Column that indicates the proband.}

\item{keyVars}{Character vector of column names for the variables in the sort key}

\item{keyVars.male}{Optional character vector of column names for the variables in the sort key
that are specific to males}

\item{keyVars.female}{Optional character vector of column names for the variables in the sort key
that are specific to females}

\item{keyWt}{Numeric vector of weights assigned to variables in the sort key, corresponding
to \code{keyVars}. If \code{NULL}, the standard deviations of the
variables in the data will be used as weights.}

\item{blockVar}{Vector of column names for the blocking variables, where families in different
blocks will not be considered when searching for duplicates.}

\item{repSN}{Number of iterations when sorting neighbors according to the sort key}

\item{windowSN}{Integer representing the size of the sliding window to use during sorted neighbors.}

\item{keyLength}{Numeric vector representing the number of key variables
(out of \code{c(keyVar.bin, keyVar.cont)}) to concatenate per sort key. If missing, the key
lengths will be randomly generated. The length of the vector should be \code{repSN}.}

\item{method}{If "intersection", we use the intersection score. If "greedy", we use the
greedy match score. If "both", we use both.}

\item{thresh}{Vector of thresholds. If \code{method = "intersection"}, then a pair is
considered neighbors if the intersection match score is greater than the threshold. If
\code{method = "greedy"}, then the threshold is treated as a percentile, and a pair is
considered neighbors if the greedy match score is greater than the percentile. If
\code{method = "both"}, then the user should provide a list, such as
\code{list(intersection = 1:7, threshold = c(0.8, 0.9))}.}

\item{priority}{A list of structure (var = 'Varx', min = TRUE) with 'Varx'
being a character value corresponding to a column in rawData.
This parameter determines how to sort the duplicates.
If \code{min = TRUE}, then we use the minimum value of 'Varx' for each duplicate entity.
Otherwise, we use the maximum value.}

\item{dateFormat}{Character string of the format of the date. This is only used
if the priority variable is a date. The format should match the formats of
class \code{POSIXlt} used in the \code{base::strptime} function.}

\item{printRuntime}{If TRUE, will print the runtime}

\item{seed}{Seed}
}
\value{
An object of class \code{\link{Duplicates}} containing the duplicate entities
and representatives for each duplicate entity (including singletons without duplicates).
}
\description{
Run entire SNIP deduplication algorithm
}
