% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortedNeighbors.R
\name{sortedNeighbors}
\alias{sortedNeighbors}
\title{Perform a single iteration of sorted neighbors.}
\usage{
sortedNeighbors(dat, key, windowSN, ID)
}
\arguments{
\item{dat}{Data frame. Data used for sorting and discovering neighbors.}

\item{key}{Integer or character vector. The columns on which to construct the sort key.}

\item{windowSN}{Integer value. Size of sliding window to use during sorted neighbrs.}

\item{ID}{Integer or character value. Identifiable column in dat.}
}
\value{
A data frame with 2 columns and n rows, each row containing neighbor pairs.
}
\description{
\code{sortedNeighbors} sorts the input data frame on the specified sort key
and returns all neighbors within a specified window.
}
