% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Truth.R
\name{Truth}
\alias{Truth}
\title{Truth object.}
\usage{
Truth(dupObj, neighborsTruth, famsTruth)
}
\arguments{
\item{dupObj}{\code{\link{Duplicates}} object.}

\item{neighborsTruth}{A data frame providing the match scores and duplication truth
outcome (true positive, true negative, false positive, false negative) for each
pair of neighbors and threshold.}

\item{famsTruth}{A data frame providing the representation status for each family
(whether they are still included in the deduplicated data) as well as the number of
extra representatives of the family.}
}
\description{
Stores the results from calling \code{\link{revealTruth}}.
}
