% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.Neighbors}
\alias{summary.Neighbors}
\alias{summary.Blocks}
\alias{summary.Scores}
\alias{summary.Duplicates}
\title{Summarize Neighbors object generic.}
\usage{
\method{summary}{Neighbors}(object, ...)

\method{summary}{Blocks}(object, ...)

\method{summary}{Scores}(object, ...)

\method{summary}{Duplicates}(object, ...)
}
\arguments{
\item{object}{\code{\link{Neighbors}} oject to be summarized.}

\item{...}{other arguments to be passed.}
}
\description{
\code{summary} summarizes the contents of the object.
}
